% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment.R
\name{augment.varest}
\alias{augment.varest}
\title{Augment a VAR model}
\usage{
\method{augment}{varest}(x, ...)
}
\arguments{
\item{...}{Additional objects to be pass through.}

\item{.x}{An object of the `varest` class.}
}
\value{
A \code{tibble}.
}
\description{
Augment for the `varest` class.
}
\examples{
model <- vars::VAR(EuStockMarkets)
augment(model)
}
