% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment.R
\name{augment_index}
\alias{augment_index}
\title{Extract the observation index from model data}
\usage{
augment_index(y)
}
\arguments{
\item{y}{Original model data.}
}
\value{
An observation index.
}
\description{
Extract the observation index from model data
}
\keyword{internal}
