% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment.R
\name{augment_matrix_long}
\alias{augment_matrix_long}
\title{Convert a model matrix to long format}
\usage{
augment_matrix_long(values, index, variable_names, value_name)
}
\arguments{
\item{values}{A numeric matrix.}

\item{index}{Observation indices.}

\item{variable_names}{Variable names.}

\item{value_name}{Name of the resulting value column.}
}
\value{
A tibble in long format.
}
\description{
Convert a model matrix to long format
}
\keyword{internal}
