% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.R
\name{autoplot.tv_fevd}
\alias{autoplot.tv_fevd}
\alias{autoplot.tv_irf}
\alias{autoplot.tv_predict}
\alias{autoplot.tv_augment}
\alias{autoplot.tv_tidy}
\alias{autoplot.tv_causality}
\alias{plot.tv_causality}
\alias{plot.tv_fevd}
\alias{plot.tv_irf}
\alias{plot.tv_predict}
\alias{plot.tv_augment}
\alias{plot.tv_tidy}
\title{Autoplot Methods for VAR Models}
\usage{
\method{autoplot}{tv_fevd}(object, ...)

\method{autoplot}{tv_irf}(object, ...)

\method{autoplot}{tv_predict}(object, ...)

\method{autoplot}{tv_augment}(object, ...)

\method{autoplot}{tv_tidy}(object, type = c("estimate", "statistic"), ...)

\method{autoplot}{tv_causality}(object, ...)

\method{autoplot}{tv_predict}(object, ...)

\method{plot}{tv_causality}(object, ...)

\method{plot}{tv_fevd}(object, ...)

\method{plot}{tv_irf}(object, ...)

\method{plot}{tv_predict}(object, ...)

\method{plot}{tv_augment}(object, ...)

\method{plot}{tv_tidy}(object, ...)
}
\arguments{
\item{object}{An object, whose class will determine the behavior of autoplot.}

\item{...}{Other arguments passed to specific methods.}

\item{type}{The column that should be plotted.}
}
\value{
A \code{ggplot2} object.
}
\description{
\code{autoplot()} uses ggplot2 to draw a plots for objects created with \code{tv_*()}
functions in a single command.
}
\examples{
library(ggplot2)

model <- vars::VAR(EuStockMarkets)

autoplot(tv_fevd(model))
autoplot(tv_irf(model))
}
