% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.R
\name{autoplot.tv_fevd}
\alias{autoplot.tv_fevd}
\title{Autoplot FEVD results}
\usage{
\method{autoplot}{tv_fevd}(object, .response = NULL, .shock = NULL, ...)
}
\arguments{
\item{object}{A `tv_fevd` object.}

\item{.response}{Optional response variables to display.}

\item{.shock}{Optional shocks to display.}

\item{...}{Additional arguments reserved for future use.}
}
\value{
A `ggplot` object.
}
\description{
Autoplot FEVD results
}
