% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm.test.R
\name{normality_test_name}
\alias{normality_test_name}
\title{Normalize normality test identifiers}
\usage{
normality_test_name(x)
}
\arguments{
\item{x}{Test name returned by `vars::normality.test()`.}
}
\value{
A character scalar.
}
\description{
Normalize normality test identifiers
}
\keyword{internal}
