% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serial.test.R
\name{serial_test_name}
\alias{serial_test_name}
\title{Normalize serial correlation test identifiers}
\usage{
serial_test_name(type)
}
\arguments{
\item{type}{Test type used by [vars::serial.test()].}
}
\value{
A character scalar.
}
\description{
Normalize serial correlation test identifiers
}
\keyword{internal}
