% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment.R
\name{tidy_augment_impl}
\alias{tidy_augment_impl}
\title{Extract augmented values from a VAR model}
\usage{
tidy_augment_impl(x)
}
\arguments{
\item{x}{A supported VAR model object.}
}
\value{
A `tv_augment` tibble.
}
\description{
Extract augmented values from a VAR model
}
\keyword{internal}
