% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causality.R
\name{tidy_causality_impl}
\alias{tidy_causality_impl}
\title{Extract tidy causality tests}
\usage{
tidy_causality_impl(x, ...)
}
\arguments{
\item{x}{A `varest` model object.}

\item{...}{Additional arguments passed to [vars::causality()].}
}
\value{
A `tv_causality` tibble.
}
\description{
Extract tidy causality tests
}
\keyword{internal}
