% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fevd.R
\name{tidy_fevd_impl}
\alias{tidy_fevd_impl}
\title{Extract tidy forecast error variance decompositions}
\usage{
tidy_fevd_impl(x, ...)
}
\arguments{
\item{x}{A supported VAR model object.}

\item{...}{Additional arguments passed to [vars::fevd()].}
}
\value{
A `tv_fevd` tibble.
}
\description{
Extract tidy forecast error variance decompositions
}
\keyword{internal}
