% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_irf.R
\name{tidy_irf}
\alias{tidy_irf}
\alias{tidy_irf.default}
\alias{tidy_irf.varest}
\title{Tidy IRF for a VAR Model}
\usage{
tidy_irf(x, ...)

\method{tidy_irf}{default}(x, ...)

\method{tidy_irf}{varest}(x, ...)
}
\arguments{
\item{...}{Additional objects to be pass through.}

\item{.x}{An object of the `varest` class.}
}
\value{
A \code{tibble}.
}
\description{
Tidyier for the `varest` class.
}
\examples{
model <- vars::VAR(EuStockMarkets)
tidy_irf(model)
}
