% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irf.R
\name{tidy_irf_component}
\alias{tidy_irf_component}
\title{Convert an impulse-response component to long format}
\usage{
tidy_irf_component(x, value_name)
}
\arguments{
\item{x}{A named list of impulse-response matrices.}

\item{value_name}{Name of the resulting value column.}
}
\value{
A tibble in long format.
}
\description{
Convert an impulse-response component to long format
}
\keyword{internal}
