% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irf.R
\name{tidy_irf_impl}
\alias{tidy_irf_impl}
\title{Extract tidy impulse-response functions}
\usage{
tidy_irf_impl(x, ...)
}
\arguments{
\item{x}{A supported model object.}

\item{...}{Additional arguments passed to [vars::irf()].}
}
\value{
A `tv_irf` tibble.
}
\description{
Extract tidy impulse-response functions
}
\keyword{internal}
