% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm.test.R
\name{tidy_multivariate_normality}
\alias{tidy_multivariate_normality}
\title{Tidy multivariate normality tests}
\usage{
tidy_multivariate_normality(tests)
}
\arguments{
\item{tests}{A named list of multivariate normality tests.}
}
\value{
A tibble.
}
\description{
Tidy multivariate normality tests
}
\keyword{internal}
