% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm.test.R
\name{tidy_normality_test}
\alias{tidy_normality_test}
\title{Extract tidy normality tests}
\usage{
tidy_normality_test(x, ...)
}
\arguments{
\item{x}{A supported VAR model object.}

\item{...}{Additional arguments passed to [vars::normality.test()].}
}
\value{
A `tv_normality_test` tibble.
}
\description{
Extract tidy normality tests
}
\keyword{internal}
