% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{tidy_predict}
\alias{tidy_predict}
\alias{tidy_predict.default}
\alias{tidy_predict.varest}
\title{Tidy Prediction of a VAR Model}
\usage{
tidy_predict(x, ...)

\method{tidy_predict}{default}(x, ...)

\method{tidy_predict}{varest}(x, ...)
}
\arguments{
\item{...}{Additional objects to be pass through.}

\item{.x}{An object of the `varest` class.}
}
\value{
A \code{tibble}.
}
\description{
Tidyier for the `varest` class.
}
\examples{
model <- vars::VAR(EuStockMarkets)
tidy_predict(model)
}
