% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serial.test.R
\name{tidy_serial_test_impl}
\alias{tidy_serial_test_impl}
\title{Extract a tidy serial correlation test}
\usage{
tidy_serial_test_impl(x, lags_pt, lags_bg, type, ...)
}
\arguments{
\item{x}{A supported VAR model object.}

\item{lags_pt}{Number of lags for Portmanteau tests.}

\item{lags_bg}{Number of lags for BG and ES tests.}

\item{type}{Serial correlation test type.}

\item{...}{Reserved for extensions.}
}
\value{
A `tv_serial_test` tibble.
}
\description{
Extract a tidy serial correlation test
}
\keyword{internal}
