% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm.test.R
\name{tidy_univariate_normality}
\alias{tidy_univariate_normality}
\title{Tidy univariate normality tests}
\usage{
tidy_univariate_normality(tests, variable_names)
}
\arguments{
\item{tests}{A list of univariate normality tests.}

\item{variable_names}{Names of the model variables.}
}
\value{
A tibble.
}
\description{
Tidy univariate normality tests
}
\keyword{internal}
