% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{tidy_varest_coefficients}
\alias{tidy_varest_coefficients}
\title{Extract tidy coefficients from a varest object}
\usage{
tidy_varest_coefficients(x)
}
\arguments{
\item{x}{A `varest` object.}
}
\value{
A `tv_tidy` tibble.
}
\description{
Extract tidy coefficients from a varest object
}
\keyword{internal}
