% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serial.test.R
\name{tv_serial_test}
\alias{tv_serial_test}
\alias{tv_serial_test.default}
\alias{tv_serial_test.varest}
\alias{tv_serial_test.vec2var}
\title{Tidy Residuals Test}
\usage{
tv_serial_test(x, ...)

\method{tv_serial_test}{default}(x, ...)

\method{tv_serial_test}{varest}(x, ...)

\method{tv_serial_test}{vec2var}(x, ...)
}
\arguments{
\item{x}{An object of the \code{varest} and \code{vec2var} class.}

\item{...}{Additional objects to be pass through.}
}
\value{
A \code{tibble}.
}
\description{
Tidier for the \code{varest} and \code{vec2var} class.
}
\examples{
model <- vars::VAR(EuStockMarkets)

tv_serial_test(model)
}
