% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{tv_tidy}
\alias{tv_tidy}
\alias{tv_tidy.default}
\alias{tv_tidy.varest}
\alias{tv_tidy.vec2var}
\title{Tidy a VAR model}
\usage{
tv_tidy(x, ...)

\method{tv_tidy}{default}(x, ...)

\method{tv_tidy}{varest}(x, ...)

\method{tv_tidy}{vec2var}(x, ...)
}
\arguments{
\item{x}{An object of the `varest` class.}

\item{...}{Additional objects to be pass through.}
}
\value{
A \code{tibble}.
}
\description{
Tidyier for the `varest` class.
}
\examples{
model <- vars::VAR(EuStockMarkets)

tv_tidy(model)
}
