% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment.R
\name{tv_augment}
\alias{tv_augment}
\alias{tv_augment.default}
\alias{tv_augment.varest}
\alias{tv_augment.vec2var}
\title{Augment a VAR model}
\usage{
tv_augment(x, ...)

\method{tv_augment}{default}(x, ...)

\method{tv_augment}{varest}(x, ...)

\method{tv_augment}{vec2var}(x, ...)
}
\arguments{
\item{x}{An object of the \code{varest} and \code{vec2var} class.}

\item{...}{Additional objects to be pass through.}
}
\value{
A \code{tibble}.
}
\description{
Augment for the \code{varest} and \code{vec2var} class.
}
\examples{
model <- vars::VAR(EuStockMarkets)

tv_augment(model)
}
