% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causality.R
\name{tv_causality}
\alias{tv_causality}
\alias{tv_causality.default}
\alias{tv_causality.varest}
\title{Tidy Causality Test}
\usage{
tv_causality(x, ...)

\method{tv_causality}{default}(x, ...)

\method{tv_causality}{varest}(x, ...)
}
\arguments{
\item{x}{An object of the \code{varest} class.}

\item{...}{Additional objects to be pass through.}
}
\value{
A \code{tibble}.
}
\description{
Tidier for the \code{varest} class.
}
\examples{
model <- vars::VAR(EuStockMarkets)

tv_causality(model)
}
