% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fevd.R
\name{tv_fevd}
\alias{tv_fevd}
\alias{tv_fevd.default}
\alias{tv_fevd.varest}
\alias{tv_fevd.vec2var}
\title{Tidy FEVD for a VAR Model}
\usage{
tv_fevd(x, ...)

\method{tv_fevd}{default}(x, ...)

\method{tv_fevd}{varest}(x, ...)

\method{tv_fevd}{vec2var}(x, ...)
}
\arguments{
\item{x}{An object of the `varest` class.}

\item{...}{Additional objects to be pass through.}
}
\value{
A \code{tibble}.
}
\description{
Tidyier for the `varest` class.
}
\examples{
model <- vars::VAR(EuStockMarkets)

tv_fevd(model)
}
