% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glance.R
\name{tv_glance}
\alias{tv_glance}
\title{Glance a VAR model}
\usage{
tv_glance(x, ...)
}
\arguments{
\item{x}{An object of the `varest` class.}

\item{...}{Additional objects to be pass through.}
}
\value{
A \code{tibble}.
}
\description{
Tidier for the \code{varest} class.
}
\examples{
model <- vars::VAR(EuStockMarkets)
tv_glance(model)
}
