% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irf.R
\name{tv_irf}
\alias{tv_irf}
\alias{tv_irf.default}
\alias{tv_irf.varest}
\alias{tv_irf.vec2var}
\alias{tv_irf.svarest}
\title{Tidy IRF for a VAR Model}
\usage{
tv_irf(x, ...)

\method{tv_irf}{default}(x, ...)

\method{tv_irf}{varest}(x, ...)

\method{tv_irf}{vec2var}(x, ...)

\method{tv_irf}{svarest}(x, ...)
}
\arguments{
\item{x}{An object of the `varest` class.}

\item{...}{Additional objects to be pass through.}
}
\value{
A \code{tibble}.
}
\description{
Tidyier for the `varest` class.
}
\examples{
model <- vars::VAR(EuStockMarkets)

tv_irf(model)
}
