% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm.test.R
\name{tv_normality_test}
\alias{tv_normality_test}
\alias{tv_normality_test.default}
\alias{tv_normality_test.varest}
\alias{tv_normality_test.vec2var}
\title{Tidy IRF for a VAR Model}
\usage{
tv_normality_test(x, ...)

\method{tv_normality_test}{default}(x, ...)

\method{tv_normality_test}{varest}(x, ...)

\method{tv_normality_test}{vec2var}(x, ...)
}
\arguments{
\item{x}{An object of the `varest` class.}

\item{...}{Additional objects to be pass through.}
}
\value{
A \code{tibble}.
}
\description{
Tidyier for the `varest` class.
}
\examples{
model <- vars::VAR(EuStockMarkets)

tv_normality_test(model)
}
