% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{tv_predict}
\alias{tv_predict}
\alias{tv_predict.default}
\alias{tv_predict.varest}
\title{Tidy Prediction of a VAR Model}
\usage{
tv_predict(x, ...)

\method{tv_predict}{default}(x, ...)

\method{tv_predict}{varest}(x, ...)
}
\arguments{
\item{x}{An object of the `varest` class.}

\item{...}{Additional objects to be pass through.}
}
\value{
A \code{tibble}.
}
\description{
Tidyier for the `varest` class.
}
\examples{
model <- vars::VAR(EuStockMarkets)
tv_predict(model)
}
