% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{tv_tidy}
\alias{tv_tidy}
\alias{tv_tidy.default}
\alias{tv_tidy.varest}
\title{Tidy a VAR Model}
\usage{
tv_tidy(x, ...)

\method{tv_tidy}{default}(x, ...)

\method{tv_tidy}{varest}(x, ...)
}
\arguments{
\item{x}{An object of the \code{varest} class.}

\item{...}{Additional objects to be pass through. Currenttly not used.}
}
\value{
A \code{tibble}.
}
\description{
Tidier for the \code{varest} class.
}
\examples{
model <- vars::VAR(EuStockMarkets)

tv_tidy(model)
}
